+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for Vehicle
 
+## Version 0.11.0
+
+* Fixed bug where using `forall ... in` and `exists ... in` would sometimes throw
+ `unification of lambdas not implemented` error.
+
 ## Version 0.10.0
 
 * Added an explicit Tensorflow-optimised backend under the new module `vehicle_lang.tensorflow`.
@@ -16,6 +21,8 @@
 
 * Fixed bug where `forall ... in` and `exists ... in` didn't evaluate properly during verification
   (introduced in v0.9.0).
+
+* Improved precision of constants in the verifier queries generated.
 
 ## Version 0.9.0